refactor(config): move config.py to config file and relevant __init__.py
